// src/app/api/models/doctor-dto.ts
/* tslint:disable */
export interface DoctorDTO {
  imageLink?: string;
  contactInfoId?: number;
  doctorIdpCode?: string;
  doctorSettingsId?: number;
  email?: string;
  firstName?: string;
  id?: number;
  dmsId?: string;
  paymentSettingsId?: number;
  phoneNumber?: number;
  practiceSince?: string;
  registerNumber?: string;
  specialization?: string;
  totalRating?: number;
}